didn't add this for some reason

+++ eval/eval_tools.py
@@ -166,7 +166,7 @@ def load_signal_files(
 
 
 def load_events(file):
-    output = np.loadtxt(file, delimiter=',', skiprows=1, usecols=(1, 2))
+    output = np.loadtxt(file, delimiter=',', skiprows=1, usecols=(1, 2)).astype(np.uint32)
     return output
